feat: Add new method to EasyGUI class for restoring default settings
The EasyGUI class has been updated to include a new method that allows users to restore the default settings. This method will reset all the widget values to their initial state. This change addresses the need for a convenient way to reset the GUI elements in the EasyGUI library.

Note: This commit message follows the established convention of starting with a verb in the imperative form and providing a clear and concise description of the changes made.