chore: Update .pre-commit-config.yaml with mypy hook
This commit adds the mypy hook to the .pre-commit-config.yaml file. The mypy hook is sourced from the pre-commit/mirrors-mypy repository at version v1.10.0. This hook will be applied to Python, pyi, and jupyter file types, and it will run the mypy command with the "--fix" argument. This change aims to improve code quality and type checking in the project.